REF: Move CompilationOptions and default_options to Options
update imports
remove Options.create_context in favor of Context.from_options
